+++ packages/editor/src/lib/editor/Editor.ts
@@ -7320,111 +7320,6 @@ export class Editor extends EventEmitter<TLEventMap> {
 	}
 
 	/**
-	 * Remove a frame.
-	 *
-	 * @param ids - Ids of the frames you wish to remove.
-	 *
-	 * @public
-	 */
-	removeFrame(ids: TLShapeId[]): this {
-		const frames = compact(
-			ids
-				.map((id) => this.getShape<TLFrameShape>(id))
-				.filter((f) => f && this.isShapeOfType<TLFrameShape>(f, 'frame'))
-		)
-		if (!frames.length) return this
-
-		const allChildren: TLShapeId[] = []
-		this.batch(() => {
-			frames.map((frame) => {
-				const children = this.getSortedChildIdsForParent(frame.id)
-				if (children.length) {
-					this.reparentShapes(children, frame.parentId, frame.index)
-					allChildren.push(...children)
-				}
-			})
-			this.setSelectedShapes(allChildren)
-			this.deleteShapes(ids)
-		})
-		return this
-	}
-
-	/**
-	 * Fit a frame to its content.
-	 *
-	 * @param id - Id of the frame you wish to fit to content.
-	 *
-	 * @public
-	 */
-	fitFrameToContent(id: TLShapeId): this {
-		const frame = this.getShape<TLFrameShape>(id)
-		if (!frame) return this
-
-		const childIds = this.getSortedChildIdsForParent(frame.id)
-		const children = compact(childIds.map((id) => this.getShape(id)))
-		if (!children.length) return this
-
-		const bounds = {
-			minX: Number.MAX_VALUE,
-			minY: Number.MAX_VALUE,
-			maxX: Number.MIN_VALUE,
-			maxY: Number.MIN_VALUE,
-		}
-		children.forEach((shape) => {
-			const geometry = this.getShapeGeometry(shape.id)
-			const points = this.getShapeLocalTransform(shape)!.applyToPoints(geometry.vertices)
-			points.forEach((point) => {
-				bounds.minX = Math.min(bounds.minX, point.x)
-				bounds.minY = Math.min(bounds.minY, point.y)
-				bounds.maxX = Math.max(bounds.maxX, point.x)
-				bounds.maxY = Math.max(bounds.maxY, point.y)
-			})
-		})
-		if (
-			bounds.minX === Number.MAX_VALUE ||
-			bounds.minY === Number.MAX_VALUE ||
-			bounds.maxX === Number.MIN_VALUE ||
-			bounds.maxY === Number.MAX_VALUE
-		)
-			return this
-
-		const padding = 50
-		const w = bounds.maxX - bounds.minX + 2 * padding
-		const h = bounds.maxY - bounds.minY + 2 * padding
-		const dx = padding - bounds.minX
-		const dy = padding - bounds.minY
-		// The shapes already perfectly fit the frame.
-		if (dx === 0 && dy === 0 && frame.props.w === w && frame.props.h === h) return this
-
-		const diff = new Vec2d(dx, dy).rot(frame.rotation)
-		this.batch(() => {
-			const changes: TLShapePartial[] = childIds.map((child) => {
-				const shape = this.getShape(child)!
-				return {
-					id: shape.id,
-					type: shape.type,
-					x: shape.x + dx,
-					y: shape.y + dy,
-				}
-			})
-
-			changes.push({
-				id: frame.id,
-				type: frame.type,
-				x: frame.x - diff.x,
-				y: frame.y - diff.y,
-				props: {
-					w,
-					h,
-				},
-			})
-
-			this.updateShapes(changes)
-		})
-		return this
-	}
-
-	/**
 	 * Update a shape using a partial of the shape.
 	 *
 	 * @example
